refactor(stories): set goals
