+++ src/uid.ts
@@ -1,4 +1,1 @@
-export default (): string =>
-  Math.random()
-    .toString(36)
-    .substring(2)
+export default (): string => Math.random().toString(36).substring(2)
